refactor: Config.setsnapshotspath()
Comprehensive and complex refactoring and also code moving.

The method Config.setSnapshotsPaths() has to much logic in it, not belonging in that method nor into its class. Code was extracted and "parked" in the tools.py module. This in consequences gives the ability to easier decouple code in SettingsDialog.saveProfile() (PR #1865).

Prepares for a new config management class (PR #1850) and refactoring the manage profiles dialog (PR #1865).

Close #1864