fix: Copy symlink as symlink in snapshot compare
When comparing snapshots, copy symlinks in the snapshots as symlinks, not as
their target.  The old behavior made the comparison not fully compare the actual
snapshot data and, even worse, could fill up your entire disk (when having
circular symlinks, e.g. a symlink to `/`, which copies `/` and eventually
copies the original symlink which points to `/` which ...).

Additionally fixes a crash when the compared snapshot(s) contain a symlink
pointing to a nonexistent target.

Fixes #1902.