feat: Prefix temp snapshot copies with `backintime_`
Back in Time will sometimes create a temp dir, to copy (parts of) a snapshot.
These have a generic name, in the form of `tmpXXXXXXXX_SNAPSHOTID`, where
`XXXXXXXX` are random characters and `SNAPSHOTID` is the snapshot id (e.g.
`20241121-112648-601`).  Only the latter gives a hint that these dirs are from
Back in Time but it's hard to tell as they're just some numbers.

This makes it tricky for the user to know if it's safe to delete these dirs if
they're left around (Back in Time will remove them if it exits cleanly but
that's not always possible, e.g. might hang if disk space gets full, like
in #1902 or when comparing too large snapshots).  Added an explicit
`backintime_` prefix to the temp dir to make their origin more clear.